+++ core/src/db/SnapshotHandlers.h
@@ -13,6 +13,7 @@
 
 #include "db/Types.h"
 #include "db/meta/FilesHolder.h"
+#include "db/snapshot/IterateHandler.h"
 #include "db/snapshot/Snapshot.h"
 #include "server/context/Context.h"
 #include "utils/Log.h"
